fix: removed unused reducer
