+++ src/redux/rootReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import createWebStorage from 'redux-persist/lib/storage/createWebStorage';
-import addressReducer from 'src/redux/addressReducer';
+import { addressReducer } from 'src/redux/slices/addressSlice';
 import authReducer from 'src/redux/authReducer';
 // ----------------------------------------------------------------------
