+++ src/redux/store.ts
@@ -4,7 +4,7 @@ import { persistStore, persistReducer } from 'redux-persist';
 import rootReducer from 'src/redux/rootReducer';
 import storage from 'redux-persist/lib/storage';
 import authReducer from 'src/redux/authReducer';
-import addressReducer from 'src/redux/addressReducer';
+import { addressReducer } from 'src/redux/slices/addressSlice';
 import api from 'src/redux/api/userAPI';
